requirements.txt
accelerate==0.24.0
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
asdfghjkl==0.1a2
asdl==0.1
async-timeout==4.0.3
attrs==23.1.0
backpack-for-pytorch==1.6.0
boltons==23.0.0
botorch==0.9.2
Brotli==1.0.9
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
conda==23.3.1
conda-libmamba-solver==23.3.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.1.1
cryptography==41.0.3
curvlinops-for-pytorch==1.2.1.dev25+g7d498c3
cycler==0.11.0
datasets==2.19.0
dill==0.3.7
docker-pycreds==0.4.0
einconv==0.1.0
einops==0.6.1
exceptiongroup==1.1.3
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.37
gmpy2==2.1.2
gpytorch==1.11
huggingface_hub==0.22.2
idna==3.4
iniconfig==2.0.0
jaxtyping==0.2.23
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
kiwisolver==1.4.5
lapeft-bayesopt==1.0
laplace-torch==0.1a2
libmambapy==1.4.2
lightning-utilities==0.9.0
linear-operator==0.5.1
mamba==1.4.2
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
multipledispatch==1.0.0
multiprocess==0.70.15
munkres==1.1.4
networkx==3.1
numpy==1.26.0
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
pathtools==0.1.2
patsy==0.5.3
peft==0.10.0
Pillow==9.4.0
pip==23.2.1
pluggy==1.3.0
ply==3.11
protobuf==4.24.3
psutil==5.9.5
pyarrow==12.0.1
pyarrow-hotfix==0.6
pycosat==0.6.4
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.12.2
pyro-api==0.1.2
pyro-ppl==1.8.6
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rdkit==2023.9.1
regex==2023.10.3
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
setuptools==68.1.2
sip==6.7.11
six==1.16.0
smmap==5.0.1
statsmodels==0.14.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.2.0
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.40.1
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
unfoldNd==0.2.0
unicodedata2==15.0.0
urllib3==2.0.4
wandb==0.15.11
wheel==0.41.2
xxhash==3.4.1
yarl==1.9.2
zstandard==0.19.0
